<type>: <task-id> - <subject> feat: modify docker configuration 1. Change dependency tool from pip to poetry 2. Move exception file to model dir
