+++ docker/redis/Dockerfile
@@ -6,7 +6,7 @@ COPY ./redis.conf /usr/local/etc/redis/redis.conf
 COPY ./functions.json /functions.json
 
 # Copy the initialization script to the container
-COPY init.sh /usr/local/bin/init.sh
+COPY ./init.sh /usr/local/bin/init.sh
 
 # Make the script executable
 RUN chmod +x /usr/local/bin/init.sh
