+++ docker/requirements.txt
@@ -1,4 +1,3 @@
-chainlit==1.3.0
 fastapi==0.111.1
 openai==1.56.2
 pglast==6.2
